Agda 2.6.4
With this Agda version there were changes [1] to instance resolution
algorithm, from agda changelog:

[Breaking] The algorithm for resolution of instance arguments has been
simplified. It will now only rely on the type of instances to
determine which candidate it should use, and no longer on their
values.

I've found this thread [2] at zulip, which might be helpful, but I
don't know how to fix the issues at felix yet.

[1] https://github.com/agda/agda/blob/master/doc/release-notes/2.6.4.md#language
[2] https://agda.zulipchat.com/#narrow/stream/238741-general/topic/Issue.20with.20instance.20search.20changes.20in.202.2E6.2E4/near/384560563