Remove old comments
Furthermore I've added explicit implicit ;) arguments to `Equivalent`
fields.  Those are actually important and it is impossible to follow
hint that was stated in the `TODO` comment, that used to said:

> TODO: move a and b arguments from methods to record.

Without those implicit args, definitions like:

```agda
∘≈ : ∀ {f g : a ⇨ b} {h k : b ⇨ c} → h ≈ k → f ≈ g → h ∘ f ≈ k ∘ g
```

wouldn't typecheck as the `Equivalent` would lock it's variables
during instantiation and we could use only those two (`a` & `b` in
this case). `_≈_` would reject morphisms `h` & `k` here as it's
objects would be different.

There is additional `TODO` comment that says:

> TODO: Replace Equivalent by Setoid?

Actually moving variables mentioned above would almost turn Equivalent
into Setoid (minus `Carrier` as a field).  Which isn't really feasible
for the reasons already mentioned.

I've took a peek how agda-categoris does it [3] and they basically put
`Equivalent` into definition of `Category` and call it _Setoid
enriched_.  I've played with that for a bit.  Initially putting `_≈_`
into lawful category.  That didn't work out however because we need
`_≈_` for defining different kinds homomorphisms, where we don't have
lawful classes in scope.  So I've moved it to the raw `Category`, but
the problem with definitions that use multiple equalities still
persists.  I think that restricting the opening of `Category` class
would be too cumbersome.  So I think this split to `Equivalent` works
nice.  I do wonder however if it's worth doing the instance search, if
most of the time one has to manually open appropriate one?

I've also updated the readme to reflect supported versions.

[3] https://github.com/agda/agda-categories/blob/master/src/Categories/Category/Core.agda#L21