Check for leftover white space
Agda stdlib team recommends to run
[fix-whitespace](https://github.com/agda/agda-stdlib/blob/master/HACKING.md#how-to-enforce-whitespace-policies)
in order to get rid of unnecessary white space characters.  I've added
this tool to the nix dev shell, as well as formed a check that will be
can be run with `nix flake check` command.  As of now this is entirely
opt-in as this check is not enabled in (and thus won't fail) the CI.