Move isEquivalence to ≈-Reasoning
Turns out agda can infer proper Equivalent typeclass from the use of
`_≈_`, based on the arguments.  What is problematic are the implicit
arguments to the `trans`, `sym` (and `refl`).  But those are used in
the proofs.  I've noticed that agda-categories [1] put those operators
inside their equivalent of `≈-Reasoning` which can (even have to) be
opened independently of `Equivalent`.  This way we can keep
`Equivalent` open in `public` and enjoy nice definitions.  What we
need to do is a little extra work of opening `≈-Reasoning` with proper
Equivalent instance as parameter.

In order to signify the change standard functions were renamed:
- refl to refl≈
- sym to sym≈
- trans to _;_

I didn't removed flipped trans (_•_) as there were some uses of it in
the code.  We can remove them separately if that'll be deemed
necessary.

[1] https://github.com/agda/agda-categories/blob/master/src/Categories/Category/Core.agda#L90